+++ lib/i18next.dart
@@ -1,6 +1,7 @@
 library i18next;
 
 export 'src/i18next.dart';
+export 'src/i18next_localization_delegate.dart';
 export 'src/options.dart';
 export 'src/resource_store.dart';
 export 'src/utils.dart';