+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Reduce description size
 * Change the namespaces type from `Map<String, Map<String, Object>> -> Map<String, Object>`
 * Adds I18Next.of(BuildContext) from Localizations
+* Adds `I18NextLocalizationDelegate`
 
 ## [0.0.1+5]