+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Adds `I18NextLocalizationDelegate`
 * Adds convenience methods to `ResourceStore` for adding, removing, and verifiying locales and namespaces
 * Adds asset bundle data source and the LocalizationDataSource interface
+* Changes links to nubank/i18next
 
 ## [0.0.1+5]