+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Adds convenience methods to `ResourceStore` for adding, removing, and verifiying locales and namespaces
 * Adds asset bundle data source and the LocalizationDataSource interface
 * Changes links to nubank/i18next
+* Adds example app
 
 ## [0.0.1+5]